Separate heartbeat client from Python heartbeat server
